Add Caching of scaled background image and mask.
Creates a BodgedReplaceBackgroundApp with two additional features, both
aimed at reducing resource, particularly for use on machines with no gpu.

First: cache the scaled background image.  The camera size isn't going to
change that frequently after all.

Second: added a basic version of only re-computing the mask ones every x
frames.  Most of the time I don't move that much, and I'd rather the mask
were a little slow to update than the image was.

Yes, this is a crude hack, but it proves the concept and that's enough for
tonight.